# server/docs/openapi.yaml
openapi: 3.0.3
...
info:
  title: Quikko API
  version: 1.0.0
  description: |
    API del acortador de URLs con métricas en tiempo real **Quikko**.

    ## Contrato de respuesta (envelope)
    Salvo el export CSV y la redirección, **toda** respuesta usa el envelope estándar:
    ```json
    { "success": true, "data": { ... }, "error": null, "meta": null }
    ```
    - `success: true` ⇒ `data` poblado, `error: null`.
    - `success: false` ⇒ `data: null`, `error` con `code` y `message`.
    - `meta` solo aparece en respuestas paginadas (se omite en el resto).

    El status HTTP sigue siendo semántico (200/201/400/401/403/404/409/410/429/500).

    ## Autenticación
    Los endpoints protegidos requieren `Authorization: Bearer <accessToken>` (JWT).
    El access token se obtiene en `register`/`login` y se renueva con `refresh`.
    El plan (free/pro) **no** viaja en el JWT: se consulta en DB en cada operación.

    ## WebSocket (tiempo real)
    `GET /ws?token=<accessToken>` hace upgrade a WebSocket. OpenAPI 3.0 no modela
    WebSockets; el protocolo (mensajes `subscribe`/`unsubscribe`/`click_event`/`error`)
    está documentado en `docs/realtime-protocol.md`.
servers:
  - url: http://localhost:8080
    description: Desarrollo local (ajustar al valor de BASE_URL si difiere)
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Registrar un usuario nuevo
      description: |
        Crea el usuario (plan Free por defecto) e inmediatamente emite un par de tokens.
        Rate limit por IP (`AUTH_RATE_LIMIT_PER_MIN`).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Usuario creado y autenticado
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthResponse'
        '400': { $ref: '#/components/responses/ValidationError' }
        '409':
          description: Email ya registrado (`AUTH_EMAIL_TAKEN`)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Iniciar sesión
      description: |
        Devuelve el mismo error (`AUTH_INVALID_CREDENTIALS`) tanto si el email no existe
        como si la contraseña es incorrecta (no filtra existencia). Rate limit por IP.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Autenticado
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthResponse'
        '400': { $ref: '#/components/responses/ValidationError' }
        '401':
          description: Credenciales inválidas (`AUTH_INVALID_CREDENTIALS`)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Renovar el access token
      description: |
        Recibe un refresh token válido y emite un nuevo access token.
        La respuesta solo trae `accessToken` (sin `refreshToken` ni `user`).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Nuevo access token emitido
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthResponse'
        '400': { $ref: '#/components/responses/ValidationError' }
        '401':
          description: Refresh token expirado (`AUTH_TOKEN_EXPIRED`) o inválido (`AUTH_TOKEN_INVALID`)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: Datos del usuario autenticado
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Usuario actual
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserDTO'
        '401': { $ref: '#/components/responses/Unauthorized' }
    delete:
      tags: [Auth]
      summary: Eliminar la propia cuenta
      description: |
        Exige la contraseña actual. Orquesta el borrado en cascada: purga el cache de
        redirección de cada URL, elimina todas las URLs del usuario en Mongo y luego el
        usuario. Los clics históricos en InfluxDB **no** se eliminan (fuera de alcance).
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteAccountRequest'
      responses:
        '200': { $ref: '#/components/responses/EmptySuccess' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401':
          description: Contraseña incorrecta (`AUTH_INVALID_CREDENTIALS`) o token inválido (`AUTH_TOKEN_INVALID`)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /api/v1/auth/me/plan:
    patch:
      tags: [Auth]
      summary: Cambiar el plan del propio usuario
      description: |
        Herramienta de prueba mientras no hay pasarela de pago: cualquier usuario
        autenticado puede cambiar su propio plan. Devuelve el usuario actualizado.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePlanRequest'
      responses:
        '200':
          description: Plan actualizado
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserDTO'
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/auth/me/password:
    patch:
      tags: [Auth]
      summary: Cambiar la contraseña
      description: Verifica la contraseña actual antes de reemplazarla.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200': { $ref: '#/components/responses/EmptySuccess' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401':
          description: Contraseña actual incorrecta (`AUTH_INVALID_CREDENTIALS`) o token inválido (`AUTH_TOKEN_INVALID`)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /api/v1/auth/me/summary:
    get:
      tags: [Auth]
      summary: Resumen de cuenta
      description: Plan, límite de URLs activas (null si Pro), conteo de activas y total de clics.
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Resumen de cuenta
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AccountSummaryResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ──────────────────────────────── URLs ────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT de acceso: `Authorization: Bearer <accessToken>`'
...
  responses:
    EmptySuccess:
      description: 'Operación exitosa sin payload (data: null)'
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Envelope'
              - type: object
                properties:
                  data:
                    description: Sin payload en esta operación (siempre null).
                    example: null
    ValidationError:
      description: Body o parámetros inválidos (`VALIDATION_ERROR`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unauthorized:
      description: Falta el token o es inválido/expirado (`AUTH_TOKEN_INVALID` / `AUTH_TOKEN_EXPIRED`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    RateLimited:
      description: Límite de peticiones excedido (`RATE_LIMIT_EXCEEDED`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    InternalError:
      description: Error interno (`INTERNAL_ERROR`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
...
    Envelope:
      type: object
      description: Sobre común de toda respuesta JSON de la API.
      required: [success, data, error]
      properties:
        success:
          type: boolean
        data:
          description: |
            Payload del endpoint (su forma se especializa por respuesta vía allOf).
            Es `null` cuando `success` es false.
        error:
          description: 'Poblado solo cuando `success` es false; `null` en caso de éxito.'
          allOf:
            - $ref: '#/components/schemas/ApiError'
        meta:
          description: Presente solo en respuestas paginadas.
          allOf:
            - $ref: '#/components/schemas/Meta'
    ErrorEnvelope:
      type: object
      description: 'Envelope de error (success: false, data: null, error poblado).'
      required: [success, data, error]
      properties:
        success: { type: boolean, example: false }
        data:
          description: Siempre null en una respuesta de error.
          example: null
        error:
          $ref: '#/components/schemas/ApiError'
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Código de error estable
          example: VALIDATION_ERROR
        message:
          type: string
          example: Cuerpo de la solicitud inválido.
    Meta:
      type: object
      description: Bloque de paginación (solo en listados).
      required: [page, limit, total, totalPages]
      properties:
        page: { type: integer, example: 1 }
        limit: { type: integer, example: 20 }
        total: { type: integer, format: int64, example: 134 }
        totalPages: { type: integer, example: 7 }

    # ── Auth ──
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8, example: superSecret1 }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken: { type: string }
    UpdatePlanRequest:
      type: object
      required: [plan]
      properties:
        plan:
          type: string
          enum: [free, pro]
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string }
        newPassword: { type: string, minLength: 8 }
    DeleteAccountRequest:
      type: object
      required: [password]
      properties:
        password: { type: string }
    AuthResponse:
      type: object
      description: |
        En register/login trae accessToken + refreshToken + user. En refresh solo
        trae accessToken (refreshToken y user se omiten).
      required: [accessToken]
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        user:
          $ref: '#/components/schemas/UserDTO'
    UserDTO:
      type: object
      required: [id, email, plan]
      properties:
        id: { type: string }
        email: { type: string, format: email }
        plan:
          type: string
          enum: [free, pro]
    AccountSummaryResponse:
      type: object
      required: [email, plan, createdAt, activeUrlsCount, totalClicks]
      properties:
        email: { type: string, format: email }
        plan:
          type: string
          enum: [free, pro]
        createdAt: { type: string, format: date-time }
        activeUrlsCount: { type: integer, format: int64 }
        activeUrlsLimit:
          type: integer
          nullable: true
          description: Límite de URLs activas del plan Free; null si Pro (ilimitado).
        totalClicks: { type: integer, format: int64 }

    # ── URLs ──
...
security:
  - bearerAuth: []